AbstractPaginator must implement Contracts interface. Must be one pagination interface in Database library Abstractions fixes - length aware context, interfaces completeness
